fix: /parser 코드 수정 및 개편 - analyzeRoutes 함수에서 ts 결과값이 기대값과 다른점을 수정하였습니다. - 수정하는 과정에서 긴코드를 나누고 정리하였습니다.
